Fix test_fitpropertybrowser unreliable test
The fitpropertybrowser was segfaulting when creating the editors within the init.
This is because the maps containing the property managers and factories are not cleared in the browser destructor as the properties can still be used elsewhere. In testing however these are not cleared and result in pointers to the managers and factories remaining even though the manager has been deleted.
This commit clears these maps after each test so it should not fail unreliably.